chore: organize kineokit, fix: kineokit couldn't import configuration file
